requirements.txt
black==21.5b2
click==8.0.1
paho-mqtt==1.5.1
w1thermsensor==2.0.0
gpiozero==1.6.2
RPi.GPIO==0.7.0
